requirements.audio.txt
fastapi==0.104.1
uvicorn==0.24.0
suno-bark==0.0.18
scipy==1.11.4
librosa==0.10.1
soundfile==0.12.1